feat: Add query to retrieve company names with price > 5000
This commit adds a new query to retrieve the names of companies from the 'PRODUCT' table where the price is greater than 5000. The query selects the 'Company' column from the 'PRODUCT' table where the price is greater than 5000. This feature will be useful for retrieving company information based on their product price.