feat: Add queries to retrieve company and customer names
This commit adds two new queries to retrieve information from the database. The first query retrieves the names of companies from the 'PRODUCT' table where the price is greater than 5000. The second query retrieves the names of customers from the city of Madras. These features will be useful for retrieving company and customer information based on specific criteria.